refactor(cypress): improve sign-in process for Opal Frontend
This commit refactors the sign-in process for the Opal Frontend. It introduces a caching mechanism using the `cy.session` function to reduce the number of requests when SSO (Single Sign-On) is enabled. This change aims to improve the performance of the test pipelines, especially when SSO is enabled. The code also handles different scenarios, such as localhost or PR environments, and Microsoft SSO login.